Move models to server folder
Includes refactor of Models class to load tables in parallel
