+++ package.json
@@ -5,6 +5,7 @@
     "start": "grunt build;node server.js"
   },
   "dependencies": {
+    "async": "^2.0.0-rc.3",
     "dotenv": "^2.0.0",
     "express": "^4.13.4",
     "express-handlebars": "^3.0.0",
